move: 타임 캡슐 상세 페이지 경로 이동 및 파일 이름 변경
